fixed scailling of recipes

+++ ST10086197_PROG2A_POE_PART1_Given/Ingredient.cs
@@ -7,7 +7,7 @@ namespace ST10086197_PROG2A_POE_PART1_Given
 	{
 
         public string ingrName { get; set; }
-        public int ingrQuantity { get; set; }
+        public double ingrQuantity { get; set; }
         public string ingrUnitOfMeasurement { get; set; }
         public int Calories { get; set; }
         public string FoodGroup { get; set; }
